Adding swift package manager manifest file Cleanup Removing subfolders to flatten source
