+++ .travis.yml
@@ -1,6 +1,6 @@
 sudo: false
 language: objective-c
-osx_image: xcode7.1
+osx_image: xcode7.2
 cache: cocoapods
 podfile: SwiftlineTests/Podfile
