+++ Source/AskSettings.swift
@@ -6,8 +6,6 @@
 //  Copyright © 2015 Omar Abdelhafith. All rights reserved.
 //
 
-import Foundation
-
 
 /// Settings used to costumize the behaviour of ask()
 public class AskSettings<T: ArgConvertibleType> {
