+++ Source/StringStyleColorizer.swift
@@ -6,9 +6,6 @@
 //  Copyright © 2015 Omar Abdelhafith. All rights reserved.
 //
 
-import Foundation
-
-
 
 public extension String {
